+++ src/utf16/convert-to-latin1.c
@@ -37,5 +37,5 @@ utf16le_convert_to_latin1(const utf16_t *data, size_t len, latin1_t *result) {
 
   if (overflow & 0xff00) return 0;
 
-  return result - start;
+  return (size_t) (result - start);
 }
