+++ src/utf8/convert-to-utf32.c
@@ -54,8 +54,8 @@ utf8_convert_to_utf32(const utf8_t *data, size_t len, utf32_t *result) {
       if ((data[pos + 1] & 0b11000000) != 0b10000000) {
         return 0;
       }
-      uint32_t code_point = ((leading_byte & 0b00011111) << 6) |
-                            (data[pos + 1] & 0b00111111);
+      uint32_t code_point = (uint32_t) (((leading_byte & 0b00011111) << 6) |
+                                        (data[pos + 1] & 0b00111111));
       if (code_point < 0x80 || 0x7ff < code_point) {
         return 0;
       }
@@ -71,9 +71,9 @@ utf8_convert_to_utf32(const utf8_t *data, size_t len, utf32_t *result) {
       if ((data[pos + 2] & 0b11000000) != 0b10000000) {
         return 0;
       }
-      uint32_t code_point = ((leading_byte & 0b00001111) << 12) |
-                            ((data[pos + 1] & 0b00111111) << 6) |
-                            (data[pos + 2] & 0b00111111);
+      uint32_t code_point = (uint32_t) (((leading_byte & 0b00001111) << 12) |
+                                        ((data[pos + 1] & 0b00111111) << 6) |
+                                        (data[pos + 2] & 0b00111111));
       if (code_point < 0x800 || 0xffff < code_point || (0xd7ff < code_point && code_point < 0xe000)) {
         return 0;
       }
@@ -92,10 +92,10 @@ utf8_convert_to_utf32(const utf8_t *data, size_t len, utf32_t *result) {
       if ((data[pos + 3] & 0b11000000) != 0b10000000) {
         return 0;
       }
-      uint32_t code_point = ((leading_byte & 0b00000111) << 18) |
-                            ((data[pos + 1] & 0b00111111) << 12) |
-                            ((data[pos + 2] & 0b00111111) << 6) |
-                            (data[pos + 3] & 0b00111111);
+      uint32_t code_point = (uint32_t) (((leading_byte & 0b00000111) << 18) |
+                                        ((data[pos + 1] & 0b00111111) << 12) |
+                                        ((data[pos + 2] & 0b00111111) << 6) |
+                                        (data[pos + 3] & 0b00111111));
       if (code_point <= 0xffff || 0x10ffff < code_point) {
         return 0;
       }
@@ -106,5 +106,5 @@ utf8_convert_to_utf32(const utf8_t *data, size_t len, utf32_t *result) {
     }
   }
 
-  return result - start;
+  return (size_t) (result - start);
 }
